client/react: make Assignee a component, use it from details ticket view
