+++ client/react/src/components/Ticket.jsx
@@ -2,6 +2,7 @@ import { useCallback, useEffect, useState } from "react";
 import { useParams } from "react-router-dom";
 import { useAuthn } from "../AuthnContext";
 import { Authz } from "@styra/opa-react";
+import Assignee from "./Assignee";
 
 export default function Ticket() {
   const { user, tenant } = useAuthn();
@@ -69,7 +70,9 @@ export default function Ticket() {
         <div id="description">{ticket.description}</div>
 
         <label htmlFor="assignee">Assignee</label>
-        <div id="assignee">{ticket.assignee || "unassigned"}</div>
+        <div id="assignee">
+          <Assignee ticket={ticket} />
+        </div>
 
         <label htmlFor="resolved">Resolved</label>
         <div id="resolved">{ticket.resolved ? "yes" : "no"}</div>
